Standardise wrapper methods (see description)
If a wrapper is for a type that has a lot of fields (e.g. ScriptParameter), having a constructor pass in all of them would be a bit absurd.

To get around this, we'll allow the user to construct these types in Lua by calling the constructor and providing values for each of the fields manually. For example;

param = ScriptParameter()
param.Name = "New"
param.Health = 1
param.Score = 100

If a type is small enough that it can be constructed inline, we'll set up "SetX" methods that take in the values of each field. Otherwise, take in a DynValue and parse the table as the target class and handle whether it should be added to the list or replaced.

Additionally, we may as well have methods that take in the DynValue for types that can be constructed inline, in case that's what the user feels like doing in the script.

If a type is also small enough that it doesn't warrant a "SetX" method, we can allow for "AddX" and "RemoveX" methods instead, say if a type has a field for a name and immediately hands off additional fields to separate arguments in a list (e.g. AssetCategory).